models: Add patch labels
Labels are arbitrary bits of metadata attached to a patch. They can
be used to signify priority, category, or other similar information.
They can also be used to filter patches and identify the ones most
interesting to a given user.

Labels can be associated with a project to ensure that one project can
use a totally different set of labels to another and to, in the future,
allow a project administrator to use their own labels. However, they can
also be global, which is useful for things that would be common across
multiple projects such as "RFC".

Signed-off-by: Stephen Finucane <stephen at that.guru>